Resolve name collisions
At the moment when two items to be saved have the same directory name, restic
only saves the first one to the repo. Let's say we have a structure like this:

    dir1
    └── subdir
        └── file
    dir2
    └── subdir
        └── file

When restic is run on `dir1/subdir` and `dir2/subdir`, it will only save the
first `subdir`:

    $ restic backup dir1/subdir dir2/subdir
    [...]

    $ restic ls latest
    drwxr-xr-x  1000   100      0 2017-08-27 20:56:39 /subdir
    -rw-r--r--  1000   100     17 2017-08-27 20:56:39 /subdir/file

That's obviously a bad thing, caused by an early decision to strip the full
path to the files/dirs to save and only leave the last directory.

This commit partly resolves this by handling colliding names and resolving the
conflicts. Restic will now append a counter (`-123`) until the conflict is
resolved. So in the example above, we'll end up with the following structure:

    $ restic ls latest
    drwxr-xr-x  1000   100      0 2017-08-27 20:56:39 /subdir
    -rw-r--r--  1000   100     17 2017-08-27 20:56:39 /subdir/file
    drwxr-xr-x  1000   100      0 2017-08-27 20:56:46 /subdir-1
    -rw-r--r--  1000   100     17 2017-08-27 20:56:46 /subdir-1/file

This partly resolves #549.